Move `test` folder under `src`
