+++ functions/tsconfig.json
@@ -12,5 +12,5 @@
   },
   "compileOnSave": true,
   "include": ["src/**/*.ts", "../src/models/Event.ts"],
-  "exclude": ["test/"]
+  "exclude": ["src/test"]
 }
